// src/types/event.ts
export interface EventData {
    id: number;
    day: string;
    month: string;
    year: string;
    date: string;
    type: string;
    title: string;
    subtitle: string;
    location: string;
    focus: string;
    description: string;
    wisdomSummary: string;
    tag: string;
    color: string;
    accent: string;
    image: string;
    number: string;
    placement: string;
}